requirements_mariadb_backup.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-Limiter==3.5.0
Flask-Login==0.6.3
mysql-connector-python==8.2.0
requests==2.31.0
gunicorn==21.2.0
python-dotenv==1.0.0
cryptography==41.0.7
stripe==7.8.0
paypalrestsdk==1.13.3